# refactor metrics calculations; improve performance of arrival history and wait times