--- a/errors.py
+++ b/errors.py
@@ -1,3 +1,0 @@
-class StopNotOnRouteError(Exception):
-   """Raised when stop does not exist on route"""
-   pass