--- a/models/errors.py
+++ b/models/errors.py
@@ -1,0 +1,5 @@
+class ValidationError(Exception):
+   pass
+
+class ArrivalHistoryNotFoundError(Exception):
+   pass