This PR initially was about improving performance for computing wait times over all stops and routes (for the isochrone map). In the process of doing that, I ended up refactoring the metrics API calculations to make it easier to maintain and add new metrics without reducing performance. 

**Refactor metrics API calculations**

This refactor allows calculating stats for wait times, trip times, headways, etc., in separate methods, while reusing the arrival history and data frames to avoid reducing performance.

The metrics API creates one RouteMetrics object for a particular route, and calls methods on that object to compute wait times, trip times, and headways for various date/time ranges. These methods separate the logic for wait times, trip times, and headways to make the code easier to maintain. These methods also allow the caller to specify which stats they want (e.g. avg, count, median, percentiles, histogram, etc) so the server doesn't do unnecessary work if the client doesn't need a particular statistic.

The RouteMetrics object caches arrival history and data frame objects so that each metric doesn't need to reload the raw arrival data. Once we add date ranges to the UI, this will become more important because it might be necessary to load >100 days of arrival data for wait times, trip times, headways, etc.

RouteMetrics also will make it straightforward to combine /metrics_by_interval and /metrics into the same API method so that the data doesn't need to be loaded twice (although not implemented in this PR). I think it would make sense to have one metrics method that allows the client to specify what metrics it wants via a GraphQL query.

**Improve arrival history performance by removing datetimes**

Profiling showed that adding datetime information to the arrival history data frame had a major impact on performance (not really noticeable in the metrics API, but noticeable when computing wait times/trip times over all stops). Since the various metrics for wait times, trip times, headways etc. only use timestamps, and datetimes are only necessary in the command line scripts when printing the raw data, I updated the arrival history data frame to work entirely on timestamps.

**Improve wait times performance and calculate exact wait time statistics**

Using numpy instead of pandas to calculate the sampled wait times made a big difference in performance when computing wait times over all stops.

I also implemented exact calculations for average wait times, quantiles and histograms (i.e., calculations that don't rely on sampling the wait times every 60 seconds) as described in the comments for wait_times.py. 

The exact methods are somewhat faster than the optimized sampled methods, although the performance difference is not very significant because both approaches can compute wait times for all stops on all routes for all displayed time intervals one day in about 20 seconds or less. The exact methods return slightly more precise wait time statistics (average wait times calculated with a sampling time of 60 seconds have a mean absolute error of about 8 seconds across all stops).

I'm still working on the script to compute and cache wait times over all stops for different time ranges, so that isn't included in this PR. I also didn't address performance of computing trip times in this PR, but planning to work on that soon.