@jtanquil 2019-06-07
Nice, the cdf error doesn't appear to be happening anymore. I'm still getting a few errors: `http://localhost:5000/metrics?route_id=S&start_stop_id=5728&end_stop_id=5730` throws a `TypeError`:
![image](https://user-images.githubusercontent.com/18497540/59081025-dc14b400-88a0-11e9-8e5c-04df49ac4ca9.png)