@youngj 2019-06-07
Good catch - timestamp was None instead of np.nan. fixed